pyproject.toml
[tool.poetry]
name = "pycon-se-bot"
version = "0.1.0"
description = ""
authors = ["Nikita Churikov <nikita@chur.ru>"]
readme = "README.md"
packages = [{include = "pycon_se_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
aiohttp = "^3.8.6"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
tortoise-orm = "^0.20.0"
aiopg = "^1.4.0"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
tortoise-orm-stubs = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = 'black'

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | venv
)/
'''

[tool.mypy]
strict = true
python_version = "3.10"
warn_return_any = true
exclude = ["venv", ".venv"]

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
